+++ manta-pay/src/test/signer.rs
@@ -26,18 +26,18 @@ use crate::{
     },
     simulation::{
         ledger::{Ledger, LedgerConnection, SharedLedger},
-        sample_signer,
+        sample_signer, sample_signer_from_seed,
     },
 };
 use alloc::sync::Arc;
 use manta_accounting::{
     transfer::{canonical::Transaction, IdentifiedAsset, Identifier},
-    wallet::{test::PublicBalanceOracle, Wallet},
+    wallet::Wallet,
 };
 use manta_crypto::{
     algebra::HasGenerator,
     arkworks::constraint::fp::Fp,
-    rand::{fuzz::Fuzz, CryptoRng, OsRng, Rand, RngCore},
+    rand::{fuzz::Fuzz, OsRng, Rand},
 };
 use manta_util::vec::VecExt;
 use std::{env, fs::OpenOptions, io::Error};
@@ -199,20 +199,18 @@ pub fn load_ledger() -> Result<SharedLedger, Error> {
 }
 
 /// Creates new wallet from `account_id`, `initial_balance`, `asset_id` and `ledger`.
-async fn create_new_wallet<R>(
+async fn create_new_wallet(
     account_id: [u8; 32],
     initial_balance: u128,
     asset_id: u128,
     ledger: SharedLedger,
-    rng: &mut R,
-) -> TestWallet
-where
-    R: CryptoRng + RngCore + ?Sized,
-{
+    seed: [u8; 32],
+) -> TestWallet {
     let directory = tempfile::tempdir().expect("Unable to generate temporary test directory.");
     let (proving_context, _, parameters, utxo_accumulator_model) =
         load_parameters(directory.path()).expect("Failed to load parameters");
-    let signer = sample_signer(&proving_context, &parameters, &utxo_accumulator_model, rng);
+    let signer =
+        sample_signer_from_seed(&proving_context, &parameters, &utxo_accumulator_model, seed);
     let asset_id = asset_id.into();
     ledger
         .write()
@@ -222,51 +220,45 @@ where
     TestWallet::new(ledger_connection, signer)
 }
 
-/// This tests that wallets preserve the invariants. After loading the [`SharedLedger`],
-/// it executes `NUMBER_OF_RUNS` times the following steps:
-/// 1) Creates a new wallet, resets it and syncs it.
-/// 2) Privatizes some tokens and syncs.
-/// 3) Sends some private tokens to another zkAddress and syncs.
-/// 4) Reclaims some tokens and syncs.
-/// 5) Restarts the wallet.
-/// 6) Checks that the public and private balances are correct.
+/// Tests that pruning is safe and doesn't delete necessary Merkle proofs.
 #[ignore] // We don't run this test on the CI because it takes a long time to run.
 #[tokio::test]
 async fn find_the_bug() {
     let mut rng = OsRng;
     let asset_id = 8;
-    println!("Loading ledger");
     let ledger = load_ledger().expect("Error loading ledger");
-    const NUMBER_OF_RUNS: usize = 100;
-    for run in 0..NUMBER_OF_RUNS {
-        println!("Run number {:?}", run + 1);
-        // 1) create new wallet, reset it and sync. (zkBalance = 0)
-        let account_id = rng.gen(); // fixed account id
-        let initial_balance = rng.gen_range(Default::default()..u32::MAX as u128);
-        println!("Initial public balance: {initial_balance:?}");
-        let mut wallet = create_new_wallet(
-            account_id,
-            initial_balance,
-            asset_id,
-            ledger.clone(),
-            &mut rng,
-        )
-        .await;
+    const NUMBER_OF_RUNS: usize = 10;
+    let account_id = rng.gen();
+    let mut public_balance = rng.gen_range(Default::default()..u32::MAX as u128);
+    let mut wallet = create_new_wallet(
+        account_id,
+        public_balance,
+        asset_id,
+        ledger.clone(),
+        rng.gen(),
+    )
+    .await;
+    let mut private_balance = 0;
+    for _ in 0..NUMBER_OF_RUNS {
+        // 1) create new wallet, reset it, sync and prune.
         wallet.reset_state();
         wallet.load_initial_state().await.expect("Sync error");
         wallet.sync().await.expect("Sync error");
-        // 2) privatize `to_mint` tokens and sync (zkBalance = to_mint)
-        let to_mint = rng.gen_range(Default::default()..initial_balance);
-        println!("To Private: {to_mint:?}");
+        wallet.signer_mut().prune();
+        // 2) privatize `to_mint` tokens, sync and prune.
+        let to_mint = rng.gen_range(Default::default()..public_balance);
+        public_balance -= to_mint;
+        private_balance += to_mint;
         let to_private = Transaction::<Config>::ToPrivate(Asset::new(asset_id.into(), to_mint));
         wallet
             .post(to_private, Default::default())
             .await
             .expect("Error posting ToPrivate");
         wallet.sync().await.expect("Sync error");
-        // 3) send `to_send` tokens to another zkAddress and sync (zkBalance = to_mint - to_send)
-        let to_send = rng.gen_range(Default::default()..to_mint);
-        println!("Private Transfer: {to_send:?}");
+        wallet.signer_mut().prune();
+        // 3) send `to_send` tokens to another zkAddress, sync and prune.
+        let to_send = rng.gen_range(Default::default()..private_balance);
+        private_balance -= to_send;
         let private_transfer =
             Transaction::<Config>::PrivateTransfer(Asset::new(asset_id.into(), to_send), rng.gen());
         wallet
@@ -274,9 +266,11 @@ async fn find_the_bug() {
             .await
             .expect("Error posting PrivateTransfer");
         wallet.sync().await.expect("Sync error");
-        // 4) reclaim `reclaim` tokens and sync (zkBalance = to_mint - to_send - reclaim)
-        let reclaim = rng.gen_range(Default::default()..to_mint - to_send);
-        println!("To Public: {reclaim:?}");
+        wallet.signer_mut().prune();
+        // 4) reclaim `reclaim` tokens, sync and prune.
+        let reclaim = rng.gen_range(Default::default()..private_balance);
+        private_balance -= reclaim;
+        public_balance += reclaim;
         let to_public =
             Transaction::<Config>::ToPublic(Asset::new(asset_id.into(), reclaim), account_id);
         wallet
@@ -284,25 +278,6 @@ async fn find_the_bug() {
             .await
             .expect("Error posting ToPublic");
         wallet.sync().await.expect("Sync error");
-        // 5: Restart wallet
-        wallet.reset_state();
-        wallet.load_initial_state().await.expect("Sync error");
-        wallet.sync().await.expect("Sync error");
-        // 6: check balances
-        let public_balance = match wallet.ledger().public_balances().await {
-            Some(asset_list) => asset_list.value(&asset_id.into()),
-            None => 0,
-        };
-        let private_balance = wallet.balance(&asset_id.into());
-        let correct_public_balance = initial_balance - to_mint + reclaim;
-        let correct_private_balance = to_mint - to_send - reclaim;
-        assert_eq!(
-            correct_private_balance, private_balance,
-            "Private balance is not correct"
-        );
-        assert_eq!(
-            correct_public_balance, public_balance,
-            "Public balance is not correct"
-        );
+        wallet.signer_mut().prune();
     }
 }